+++ profile.py
@@ -9,7 +9,9 @@ import shutil
 
 
 def profile():
-    shutil.rmtree(config.TF_TRASH_PATH)
+    if os.path.isdir(config.TF_TRASH_PATH)
+        shutil.rmtree(config.TF_TRASH_PATH)
+
     os.makedirs(config.TF_TRASH_PATH)
 
     sn = SmallNorb.from_cache()
